+++ src/lib/form/index.ts
@@ -15,6 +15,10 @@ export {
     useOnSubmitNewWithGroups,
 } from './useOnSubmitWithGroups'
 export { modelFormSchemas } from './modelFormSchemas'
+export {
+    hasUnsavedDataElements,
+    hasUnsavedTrackedEntityAttributes,
+} from './hasUnsavedDataElements'
 export * from './sectionedForm'
 export * from './useCreateModel'
 export * from './createFormError'