+++ CHANGELOG.md
@@ -1,3 +1,45 @@
+# [0.172.0](https://github.com/dhis2/metadata-management-app/compare/v0.171.1...v0.172.0) (2026-07-29)
+
+
+### Bug Fixes
+
+* fix text in unsaved notice box ([6eb18a2](https://github.com/dhis2/metadata-management-app/commit/6eb18a223747f9b897eb3f98c08a95d4591a8783))
+* footer drawer elements alignment ([cb61758](https://github.com/dhis2/metadata-management-app/commit/cb61758bf284824efe467ff6037f9c89cf91c468))
+
+
+### Features
+
+* add unsaved attributes warning for programs ([c3092cd](https://github.com/dhis2/metadata-management-app/commit/c3092cdce4612b726cfa5d7f3083f387c6056ae2))
+* unsaved element notice box ([a656ba1](https://github.com/dhis2/metadata-management-app/commit/a656ba12098a0d4656f12313fd8dc5605da897dd))
+* unsaved elements for custom and section form ([7a92517](https://github.com/dhis2/metadata-management-app/commit/7a925174543c5dc0ff03da47358316fc75cc107f))
+* unsaved elements in program stages and data sets ([b4e9bb3](https://github.com/dhis2/metadata-management-app/commit/b4e9bb368338b6f1fc3c84aad4b137b3aab2537b))
+
+## [0.171.1](https://github.com/dhis2/metadata-management-app/compare/v0.171.0...v0.171.1) (2026-07-23)
+
+
+### Bug Fixes
+
+* predictor edit issue ([#1010](https://github.com/dhis2/metadata-management-app/issues/1010)) ([5fdba50](https://github.com/dhis2/metadata-management-app/commit/5fdba50d0839764db0a4bd8663a88892bd8c0e6b))
+
+# [0.171.0](https://github.com/dhis2/metadata-management-app/compare/v0.170.2...v0.171.0) (2026-07-23)
+
+
+### Bug Fixes
+
+* custom form event propagation and footer layout ([#1005](https://github.com/dhis2/metadata-management-app/issues/1005)) ([b29a2cf](https://github.com/dhis2/metadata-management-app/commit/b29a2cf84fff6f8a10f11afe0eba5dbe2f244653))
+
+
+### Features
+
+* cat option combo merge ([#1004](https://github.com/dhis2/metadata-management-app/issues/1004)) ([b4d2f6d](https://github.com/dhis2/metadata-management-app/commit/b4d2f6d819823d621d36c43be3f8c0949631b2d6))
+
+## [0.170.2](https://github.com/dhis2/metadata-management-app/compare/v0.170.1...v0.170.2) (2026-07-23)
+
+
+### Bug Fixes
+
+* require organisation unit levels for predictors [DHIS2-12090] ([#1001](https://github.com/dhis2/metadata-management-app/issues/1001)) ([919a09a](https://github.com/dhis2/metadata-management-app/commit/919a09a77d5ca1503ca441e860ee65b6718e6ee6))
+
 ## [0.170.1](https://github.com/dhis2/metadata-management-app/compare/v0.170.0...v0.170.1) (2026-07-23)